take CollectibleActions as a shared component.
