+++ playgrounds/react-vite/src/components/collectible/index.ts
@@ -1,4 +1,3 @@
-export { Actions } from './Actions';
 export { CollectibleDetails } from './CollectibleDetails';
 export {
 	ActivitiesTable,
