+++ playgrounds/shared/src/components/collectible/index.ts
@@ -1,0 +1,1 @@
+export * from './Actions';
