+++ playgrounds/shared/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './settings';
 export * from './navigation';
+export * from './collectible';
